requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r ./base.txt

gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
psycopg2==2.8.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
Collectfast==2.1.0  # https://github.com/antonagestam/collectfast
sentry-sdk==0.14.2  # https://github.com/getsentry/sentry-python

# Django
# ------------------------------------------------------------------------------
django-storages[google]==1.9.1  # https://github.com/jschneier/django-storages
django-anymail[sendgrid]==7.0.0  # https://github.com/anymail/django-anymail

celery==4.4.0
django-celery==3.3.1
django-redis=4.11.0
redis==3.4.1
